+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.6.0 > x.x.x
 
+-   Add copyts, start_at_zero and use_wallclock_as_timestamps option
 -   Add heuristic to find core address if UI is proxied
 -   Add Turkish translation (thanks to Ramazan Sancar) ([#22](https://github.com/datarhei/restreamer-ui/issues/22))
 -   Add Danish translation (Thanks to Filip Stadler and Info)